Audit LOW batch 6: one tree-materializer, verified source adoption
- hoist the substituter's two-pass NAR unpacker into Nix.Store as the
  codebase's single tree-materializer; Eval/IO's single-pass copy
  (which typed a symlink by probing its not-yet-materialized target)
  is gone, delegating instead; the substituter re-exports for its
  callers and tests
- materializeEvalSources moves to Nix.Store and verifies adoption:
  an on-disk tree must reproduce the store path it is registered
  under (recursive NAR digest + name), or it is cleared and re-copied
  - a partial tree from an interrupted copy no longer registers
